PDFBOX-2852: remove unnecessary type casting and type arguments, as suggested by Lorenz Pahl
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1760234 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/interactive/form/PDAcroForm.java
@@ -243,7 +243,7 @@ public final class PDAcroForm implements COSObjectable
                     	}
                     	Integer pageRef = annotationToPageRef.get(widget.getCOSObject());
                     	if (pageRef != null) {
-                    		page = document.getPage((int) pageRef);
+                    		page = document.getPage(pageRef);
                     	}
                     }
